added Event Seeder migrations , etc and new test case

+++ database/migrations/2018_04_04_111930_create_venues_table.php
@@ -20,8 +20,8 @@ class CreateVenuesTable extends Migration
             $table->string('city');
             $table->string('state');
             $table->string('zip');
-            $table->float('longitude');
-            $table->float('latitude');
+            $table->decimal('longitude',11,8);
+            $table->decimal('latitude',10,8);
             $table->timestamps();
         });
     }
